requirements (2).txt
Flask
Flask-Cors
pandas
matplotlib
numpy
scipy
biopython
torch
gunicorn
